Moved and rearranged folder structure.
